auth

+++ repository/contacts.js
@@ -1,31 +1,56 @@
 const Contact = require('../models/contact');
 
-const getAll = async () => {
-    const result = await Contact.find();
+const getAll = async ({limit, page, sort}, filter, user) => {
+    let option = {};
+    if (filter) {
+        option = { owner: user.id, favorite: filter };
+    }
+    if (!filter) {
+        option = { owner: user.id };
+    }
+
+    const { docs: contacts, ...rest } = await Contact.paginate(
+        option,
+        {limit, page, sort, select: 'name email phone favorite'},
+    )
+    return { contacts, ...rest }
+}
+
+const getById = async (contactId, user) => {
+    const result = await Contact.findOne({ _id: contactId, owner: user.id }).populate({
+        path: 'owner',
+        select: 'email subscription',
+    }); 
     return result
 }
 
-const getById = async (contactId) => {
-    const result = await Contact.findOne({_id: contactId}); // Contact.findById(contactId)
+const remove = async (contactId, user) => {
+    const result = await Contact.findOneAndRemove({_id: contactId, owner: user.id}); 
     return result
 }
 
-const remove = async (contactId) => {
-    const result = await Contact.findOneAndRemove({_id: contactId}); 
+const create = async (body, user) => {
+    const result = await Contact.create({ ...body, owner: user.id }); 
     return result
 }
 
-const create = async (body) => {
-    const result = await Contact.create(body);  
-    return result
-}
-
-const update = async (contactId, body) => {
+const update = async (contactId, body, user) => {
     const result = await Contact.findOneAndUpdate(
-      { _id: contactId },
+      { _id: contactId, owner: user.id },
       { ...body },
       { new: true },
     );
+    return result
+}
+
+const checkContact = async (body, user) => {
+    const result = await Contact.findOne({
+        $or: [
+            { name: body.name },
+            { email: body.email },
+            { phone: body.phone }
+        ], owner: user.id
+    }); 
     return result
 }
 
@@ -34,5 +59,6 @@ module.exports = {
     getById,
     remove,
     create,
-    update
+    update,
+    checkContact
 }
